fpc1020: Use atomic operations instead to avoid race
The main motive was to prevent simultaneous R/W to the wakeup_enabled
variable which would then result in a undesired race condition. Memory
barriers were not created for that purpose and using them does not
guarantee a fix. Atomic operations were what we needed to serve that
purpose. In the newer fpc1020 drivers from MiCode, wakeup_enabled was
too, converted to an atomic variable.

Change-Id: I0abf3841ef54e9214a48d9f7e4f5d2c09acb63cc